Separate plugin loader
