+++ maubot/loader/__init__.py
@@ -1,0 +1,2 @@
+from .abc import PluginLoader, PluginClass
+from .zip import ZippedPluginLoader
